MoPS-package {MoPS}	R Documentation

MoPS - Model-based Periodicity Screening

Description

The MoPS package implements the model-based periodicity screen as used in Eser et al.(Mol Syst Biol, 2014) for the detection and characterization of periodic genes.

Details

Package: 	 MoPS

Type: 	 Package

Version: 	 0.99.0

Date: 	 2014-04-15

License: 	 GPL-3

MoPS provides methods for screening numerical time series data for periodicity. See the package vignette for a detailed description of the methods and recommended workflows.

Author(s)

Philipp Eser, Achim Tresch

Maintainer: Philipp Eser <eser@genzentrum.lmu.de>

References

Periodic mRNA synthesis and degradation co-operate during cell cycle gene expression (Eser et al. Mol Sys Biol, 2014)

[Package MoPS version 1.14.0 Index]