result.as.dataframe {MoPS}	R Documentation

Parameters of fitted periodic time courses.

Description

Converts the result list derived by MoPS function fit.periodic() to a data.frame.

Usage

result.as.dataframe(result.list)

Arguments

result.list
List of best fitting parameters returned by fit.periodic().

Details

This function takes as input the result list from MoPS function fit.periodic() and extracts the time course specific optimal parameters.

Value

data.frame containing the best fitting periodic parameters for each time series (rows):

columns: ID : unique identifier score : log-likelihood for periodic behaviour phi : phase lambda : period length sigma : attenuation of the signal along the complete time series mean : mean amplitude : amplitude

Author(s)

Philipp Eser, Achim Tresch

Examples

y = 2*sin(seq(0,6*pi,length.out=50)+rnorm(50))
res = fit.periodic(y)
result.as.dataframe(res)

[Package MoPS version 1.14.0 Index]